Serveis de gestor a falta de provar

+++ src/main/java/es/uji/ei102720gmtp/SanaProject/controller/ServeiController.java
@@ -1,12 +1,8 @@
 package es.uji.ei102720gmtp.SanaProject.controller;
 
-import es.uji.ei102720gmtp.SanaProject.Validation.EspaiPublicValidator;
 import es.uji.ei102720gmtp.SanaProject.Validation.ServeiEstacionalValidator;
 import es.uji.ei102720gmtp.SanaProject.Validation.ServeiPermanentValidator;
-import es.uji.ei102720gmtp.SanaProject.dao.EspaiPublicDao;
-import es.uji.ei102720gmtp.SanaProject.dao.ServeiEstacionalDao;
-import es.uji.ei102720gmtp.SanaProject.dao.ServeiInstalatEspaiDao;
-import es.uji.ei102720gmtp.SanaProject.dao.ServeiPermanentDao;
+import es.uji.ei102720gmtp.SanaProject.dao.*;
 import es.uji.ei102720gmtp.SanaProject.model.*;
 import es.uji.ei102720gmtp.SanaProject.services.InterfaceServeiService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -17,6 +13,7 @@ import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -30,6 +27,7 @@ public class ServeiController {
     private InterfaceServeiService serveiService;
     private EspaiPublicDao espaiPublicDao;
     private ServeiInstalatEspaiDao serveiInstalatEspaiDao;
+    private PeriodeServeiEspaiDao periodeServeiEspaiDao;
 
     @Autowired
     public void setServeiInstalatEspaiDao(ServeiInstalatEspaiDao serveiInstalatEspaiDao) {
@@ -54,6 +52,11 @@ public class ServeiController {
     @Autowired
     public void setEspaiPublicDao(EspaiPublicDao espaiPublicDao) {
         this.espaiPublicDao = espaiPublicDao;
+    }
+
+    @Autowired
+    public void setPeriodeServeiEspaiDao(PeriodeServeiEspaiDao periodeServeiEspaiDao) {
+        this.periodeServeiEspaiDao = periodeServeiEspaiDao;
     }
 
     @RequestMapping("/seleccionarTipusServei")
@@ -103,8 +106,15 @@ public class ServeiController {
     }
 
     @RequestMapping(value = "/deleteServeiPermanent/{nom}")
-    public String deleteServeiPermanent(Model model, @PathVariable String nom){
+    public String deleteServeiPermanent(Model model, @PathVariable String nom, RedirectAttributes redirectAttributes){
+        if (serveiService.isPermanentServeiUsed(nom)){
+            String msg = "No es pot borrar el servei " + nom + " perque s'esta usant en alguna area";
+            redirectAttributes.addFlashAttribute("alert", msg);
+            return "redirect:../listServeiPermanent";
+        }
         serveiPermanentDao.deleteServeiPermanent(nom);
+        String msg = "S'ha borrat el servei: " + nom;
+        redirectAttributes.addFlashAttribute("alert", msg);
         return "redirect:../listServeiPermanent";
     }
 
@@ -150,8 +160,16 @@ public class ServeiController {
     }
 
     @RequestMapping(value = "/deleteServeiEstacional/{nom}")
-    public String deleteServeiEstacional(Model model, @PathVariable String nom){
+    public String deleteServeiEstacional(Model model, @PathVariable String nom, RedirectAttributes redirectAttributes){
+        if (serveiService.isEstacionalServeiUsed(nom)){
+            String msg = "No es pot borrar el servei " + nom + " perque s'esta usant en alguna area";
+            redirectAttributes.addFlashAttribute("alert", msg);
+            return "redirect:../listServeiEstacional";
+        }
+
         serveiEstacionalDao.deleteServeiEstacional(nom);
+        String msg = "S'ha borrat el servei: " + nom;
+        redirectAttributes.addFlashAttribute("alert", msg);
         return "redirect:../listServeiEstacional";
     }
 
@@ -170,12 +188,14 @@ public class ServeiController {
         return "servei/listServeisEspai";
     }
 
+    // Serveis Permanents
+
     @RequestMapping(value = "/addServeiPermanentGestor/{id}")
     public String addServeiPermanentGestor(Model model, @PathVariable int id) {
         EspaiPublic espai = espaiPublicDao.getEspaiPublic(id);
 
         List<ServeiPermanent> instalats = serveiPermanentDao.getServeisPermanentsFromEspai(id);
-        List<ServeiPermanent> list = serveiService.getServeisRestants(instalats);
+        List<ServeiPermanent> list = serveiService.getServeisPermanentsRestants(instalats);
 
         model.addAttribute("espai", espai);
         model.addAttribute("serveis", list);
@@ -215,6 +235,62 @@ public class ServeiController {
         return "servei/listServeisEspai";
     }
 
+    // Serveis estacionals
 
+    @RequestMapping(value = "/addServeiEstacionalGestor/{id}")
+    public String addServeiEstacionalGestor(Model model, @PathVariable int id) {
+        EspaiPublic espai = espaiPublicDao.getEspaiPublic(id);
+
+        List<ServeiEstacional> instalats = serveiEstacionalDao.getServeisEstacionalsFromEspai(id);
+        List<ServeiEstacional> list = serveiService.getServeisEstacionalsRestants(instalats);
+
+        model.addAttribute("espai", espai);
+        model.addAttribute("serveis", list);
+
+        ServeiEstacionalComplet serveiEstacionalComplet = new ServeiEstacionalComplet();
+        model.addAttribute("serveiEstacionalComplet", serveiEstacionalComplet);
+
+        return "servei/addServeiEstacionalGestor";
+    }
+
+    @RequestMapping(value = "/addServeiEstacionalGestor", method= RequestMethod.POST)
+    public String addServeiEstacionalGestorPost(@ModelAttribute("serveiEstacionalComplet") ServeiEstacionalComplet serveiEstacionalComplet, BindingResult bindingResult, Model model) {
+        EspaiPublic espai = espaiPublicDao.getEspaiPublic(serveiEstacionalComplet.getIdEspai());
+
+
+
+        PeriodeServeiEspai periodeServeiEspai = new PeriodeServeiEspai();
+        periodeServeiEspai.setNomServei(serveiEstacionalComplet.getNom());
+        periodeServeiEspai.setIdEspai(serveiEstacionalComplet.getIdEspai());
+        periodeServeiEspai.setHoraInici(serveiEstacionalComplet.getHoraInici());
+        periodeServeiEspai.setHoraFinal(serveiEstacionalComplet.getHoraFinal());
+        periodeServeiEspai.setDataInici(serveiEstacionalComplet.getDataInici());
+        periodeServeiEspai.setDataFinal(serveiEstacionalComplet.getDataFinal());
+        periodeServeiEspaiDao.addPeriodeServeiEspai(periodeServeiEspai);
+
+        List<ServeiPermanentComplet> serveisPermanentsInstalats = serveiService.getServeiPermanentInstalats(espai.getId());
+        List<ServeiEstacionalComplet> serveiEstacionalInstalats = serveiService.getServeisEstacionalsInstalats(espai.getId());
+
+        model.addAttribute("serveisPermanents", serveisPermanentsInstalats);
+        model.addAttribute("serveisEstacionals", serveiEstacionalInstalats);
+
+        model.addAttribute("espai", espai);
+
+        return "servei/listServeisEspai";
+    }
+
+    @RequestMapping(value = "/deleteServeiEstacionalGestor/{id}/{nom}")
+    public String deleteServeiEstacionalGestor(Model model, @PathVariable int id, @PathVariable String nom){
+        serveiInstalatEspaiDao.deleteServeiInstalatEspai(id, nom);
+
+        List<ServeiPermanentComplet> serveisPermanentsInstalats = serveiService.getServeiPermanentInstalats(id);
+        List<ServeiEstacionalComplet> serveiEstacionalInstalats = serveiService.getServeisEstacionalsInstalats(id);
+
+        model.addAttribute("serveisPermanents", serveisPermanentsInstalats);
+        model.addAttribute("serveisEstacionals", serveiEstacionalInstalats);
+
+        model.addAttribute("espai", espaiPublicDao.getEspaiPublic(id));
+        return "servei/listServeisEspai";
+    }
 
 }
